rename nav to navbar
